+++ CHANGELOG.rst
@@ -24,6 +24,9 @@ Added
 Changed
 -------
 - Unknown sections in markdown format (NLU data) are not ignored anymore, but instead an error is raised.
+- Serialise ``DialogueStateTracker`` as json instead of pickle. **DEPRECATION warning**:
+  Deserialisation of pickled trackers will be deprecated in a future version. For now,
+  trackers loaded from pickle will be dumped as json.
 
 Removed
 -------